fonction move position OK, behavior1 OK

+++ POC/MPLAB/marvin_v02.X/Servo.h
@@ -18,7 +18,7 @@
 void    marvin_attach_servo(m_servo *servo, u32 *pin ,u32 *ocrs, u16 min, u16 max, u8 oc_timer, u16 periode);
 void    marvin_move_servo(m_servo *servo, s16 angle);
 void    marvin_move_servo_speed(m_servo *servo, s16 angle, u16 deg_per_periode, u16 periode_msec);
-void    marvin_move_to_position(u8 pitch_angle, u8 yaw_angle, u8 max_periode_msec);
+void    marvin_move_to_position(u8 pitch_angle, u8 yaw_angle, u16 max_periode_msec);
 
 #endif	/* SERVO_H */
